refactor(talktype): create mutations folder for talktype crud
